[#42] refactor: post like 패키지 구조 수정
